Draft: split offset from Draft.py
.
